+++ src/optlang/matrix_interface.py
@@ -1,6 +1,3 @@
-# Copyright 2013 Novo Nordisk Foundation Center for Biosustainability,
-# Technical University of Denmark.
-#
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -13,80 +10,99 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Solver interface for the OSQP solver.
+"""Generic solver interface for (sparse) matrix solvers.
 
-Wraps the OSQP solver by subclassing and extending :class:`Model`,
-:class:`Variable`, and :class:`Constraint` from :mod:`interface`.
+This interface is only supposed to be a basis to add matrix solvers to optlang. It
+creates a thin layer between all optlang operations and the standard formulation of
+LP, QP, and MILP problems.
 
-To use this interface, install the OSQP solver and the bundled python
-interface.
-Make sure that 'import osqp' runs without error.
+To create a solver from this, you need to do the following:
+
+1. Create a class that inherits from MatrixProblem and implement the `solve` method
+   and optionally overwrite the `reset` and `still_valid` methods.
+
+2. Create derived classes for Variable, Constraint, Objective, and Configuration.
+   You don't need to overwrite anything. Those should work out of the box.
+
+3. Create a derived class for Model and set the `ProblemClass` attribute to the class
+   from (1) and the `status_map` to map the solver status to an Optlang status from
+   `optlang.interface`.
+
+For an example take a look at `optlang.osqp_interface`.
 """
+
+import abc
 import logging
-import six
 import pickle
-import numpy as np
-from numpy import array, concatenate, Infinity, zeros, isnan
+from collections import defaultdict
+from typing import NamedTuple
 
-from optlang import interface, symbolics, available_solvers
-from optlang.util import inheritdocstring
-from optlang.expression_parsing import parse_optimization_expression
-from optlang.exceptions import SolverError
-
+from numpy import Infinity, array, concatenate, zeros
 from scipy.sparse import csc_matrix
 
+from optlang import available_solvers, interface, symbolics
+from optlang.exceptions import SolverError
+from optlang.expression_parsing import parse_optimization_expression
 from optlang.symbolics import add, mul
 
 log = logging.getLogger(__name__)
 
-try:
-    import osqp as osqp
-except ImportError:
-    try:
-        import cuosqp as osqp
+_STATUS_MAP = defaultdict(lambda: interface.UNDEFINED)
 
-        log.warning(
-            "cuOSQP is still experimental and may not work for all problems. "
-            "It may not converge as fast as the normal osqp package or not "
-            "converge at all."
-        )
-    except ImportError:
-        raise ImportError("The osqp_interface requires osqp or cuosqp!")
+_LP_METHODS = ("auto", )
+
+_QP_METHODS = ("auto", )
+
+_TYPES = ("continuous", "binary", "integer")
 
 
-_STATUS_MAP = {
-    "interrupted by user": interface.ABORTED,
-    "run time limit reached": interface.TIME_LIMIT,
-    "feasible": interface.FEASIBLE,
-    "primal infeasible": interface.INFEASIBLE,
-    "dual infeasible": interface.INFEASIBLE,
-    "primal infeasible inaccurate": interface.INFEASIBLE,
-    "dual infeasible inaccurate": interface.INFEASIBLE,
-    "solved inaccurate": interface.NUMERIC,
-    "solved": interface.OPTIMAL,
-    "maximum iterations reached": interface.ITERATION_LIMIT,
-    "unsolved": interface.SPECIAL,
-    "problem non convex": interface.SPECIAL,
-    "non-existing-status": "Here for testing that missing statuses are handled.",
-}
+class SparseProblem(NamedTuple):
+    """A representation of the convex optimizatgion problem in standard form.
 
-_LP_METHODS = ("auto", "primal")
+    This defines the problem in the form.
 
-_QP_METHODS = ("auto", "primal")
+    ..math::
 
-_TYPES = ("continuous",)
+      \text{minimize }& \frac{1}{2}x^T\mathbf{P}x + q^Tx \\
+      \text{s.t.: }& bounds_{.0} \leq \mathbf{A} \leq bounds_{.1} \\
+      & vbounds_{.0} \leq x \leq vbounds_{.1} \\
+      & \{x_k \in \mathbb{N}^0 \text{ if } integer_k = 1 \}
+
+    Attributes
+    ----------
+    P : csc_matrix
+        A semidefinite positive matrix specifying the coefficients for the quadratic
+        objective.
+    q : array
+        A vector specfiying the linear objective coefficients.
+    A : csc_matrix
+        The constraint matrix.
+    bounds : array
+        The lower and upper bounds corresponding to the rows in `A`.
+    vbounds : array
+        The lower and upper bounds of the variables `x`.
+    integer : array
+        Indicator vector denoting if `x[k]` is integer.
+    """
+    P : csc_matrix
+    q : array
+    A : csc_matrix
+    bounds : array
+    vbounds : array
+    integer : array
 
 
-class OSQPProblem(object):
-    """A concise representation of an OSQP problem.
+class MatrixProblem(abc.ABC):
+    """A concise representation of an LP/QP/MILP problem.
 
-    OSQP assumes that the problem will be pretty much immutable. This is
+    This class assumes that the problem will be pretty much immutable. This is
     a small intermediate layer based on dictionaries that is fast to modify
-    but can also be converted to an OSQP problem without too much hassle.
+    but can also be converted to a matrix formulation quickly.
     """
 
-    def __init__(self):
+    def __init__(self, **kwargs):
         self.variables = set()
+        self.integer_vars = set()
         self.constraints = set()
         self.constraint_coefs = dict()
         self.constraint_lbs = dict()
@@ -103,25 +119,52 @@ class OSQPProblem(object):
         self.direction = -1
         self.info = None
         self.status = None
-        self.settings = {
-            "linsys_solver": "qdldl",
-            "max_iter": 100000,
-            "eps_abs": 1e-6,
-            "eps_rel": 1e-6,
-            "eps_prim_inf": 1e-6,
-            "eps_dual_inf": 1e-6,
-            "polish": True,
-            "verbose": False,
-            "scaling": 0,
-            "time_limit": 0,
-            "adaptive_rho": True,
-            "rho": 1.0,
-            "alpha": 1.6,
-        }
-        self.__solution = None
+        self._solution = None
+        self.default_settings()
 
-    def build(self):
-        """Build the problem instance."""
+    def default_settings(self):
+        """Set the default settings of the solver.
+
+        Returns
+        -------
+        dict
+            The default settings for the solver.
+        """
+        self.settings = {
+            "presolve": True,
+            "lp_method": "auto",
+            "qp_method": "auto",
+            "time_limit": 0,
+            "primal_inf_tolerance": 1e-6,
+            "dual_inf_tolerance": 1e-6,
+            "optimality_tolerance": 1e-6,
+            "mip_tolerance": 1e-6,
+            "verbose": 1,
+            "threads": 1,
+            "iteration_limit": 100000
+        }
+
+    def build(self, add_variable_constraints=False):
+        """Build the problem instance.
+
+        Parameters
+        ----------
+        add_variable_constraints : bool
+            Whether to add constraints for each individual variables as a (sparse)
+            block diagonal to the constraint matrix A.
+
+        Returns
+        -------
+        tuple of [P, q, A, bounds, vbounds, integer] with the following components:
+            - P : Quadratic objective coefficients as an upper triangular CSC matrix
+            - q : Linear objective coefficients as an array.
+            - A : Constraint matrix as a CSC matrix.
+            - bounds : Constraint lower and upper bounds as a 2xNc matrix.
+            - vbounds : Variable bounds as a 2xNv matrix.
+            - integer : Binary array indicating whether variable i is integer
+                        (for MIPs).
+
+        """
         d = self.direction
         vmap = dict(zip(self.variables, range(len(self.variables))))
         nv = len(self.variables)
@@ -131,7 +174,7 @@ class OSQPProblem(object):
             P = array(
                 [
                     [vmap[vn[0]], vmap[vn[1]], coef * d * 2.0]
-                    for vn, coef in six.iteritems(self.obj_quadratic_coefs)
+                    for vn, coef in self.obj_quadratic_coefs.items()
                 ]
             )
             P = csc_matrix(
@@ -145,15 +188,15 @@ class OSQPProblem(object):
             self.obj_linear_coefs.values()
         )
         q = q * d
-        Av = array([[vmap[k] + nc, vmap[k], 1.0] for k in self.variables])
         vbounds = array(
             [[self.variable_lbs[vn], self.variable_ubs[vn]] for vn in self.variables]
         )
+
         if len(self.constraint_coefs) > 0:
             A = array(
                 [
                     [cmap[vn[0]], vmap[vn[1]], coef]
-                    for vn, coef in six.iteritems(self.constraint_coefs)
+                    for vn, coef in self.constraint_coefs.items()
                 ]
             )
             bounds = array(
@@ -162,157 +205,194 @@ class OSQPProblem(object):
                     for cn in self.constraints
                 ]
             )
-            A = concatenate((A, Av))
-            bounds = concatenate((bounds, vbounds))
-        else:
-            A = Av
-            bounds = vbounds
-        if A.shape[0] == 0:
-            A = None
-        else:
+            if add_variable_constraints:
+                Av = array([[vmap[k] + nc, vmap[k], 1.0] for k in self.variables])
+                A = concatenate((A, Av))
+                bounds = concatenate((bounds, vbounds))
             A = csc_matrix(
                 (A[:, 2], (A[:, 0].astype("int64"), A[:, 1].astype("int64"))),
-                shape=(nc + nv, nv),
+                shape=(bounds.shape[0], nv),
             )
-        return P, q, A, bounds
-
-    def solve(self):
-        """Solve the OSQP problem."""
-        settings = self.settings.copy()
-        P, q, A, bounds = self.build()
-        solver = osqp.OSQP()
-        if P is None:
-            # see https://github.com/cvxgrp/cvxpy/issues/898
-            settings.update({"adaptive_rho": 0, "rho": 1.0, "alpha": 1.0})
-        solver.setup(P=P, q=q, A=A, l=bounds[:, 0], u=bounds[:, 1], **settings)  # noqa
-        if self.__solution is not None:
-            if self.still_valid(A, bounds):
-                solver.warm_start(x=self.__solution["x"], y=self.__solution["y"])
-                solver.update_settings(rho=self.__solution["rho"])
-        solution = solver.solve()
-        nc = len(self.constraints)
-        nv = len(self.variables)
-        if not solution.x[0] is None:
-            self.primals = dict(zip(self.variables, solution.x))
-            self.vduals = dict(zip(self.variables, solution.y[nc : (nc + nv)]))
-            if nc > 0:
-                self.cprimals = dict(zip(self.constraints, A.dot(solution.x)[0:nc]))
-                self.duals = dict(zip(self.constraints, solution.y[0:nc]))
-        if not isnan(solution.info.obj_val):
-            self.obj_value = solution.info.obj_val * self.direction
-            self.status = solution.info.status
+        elif add_variable_constraints:
+            Av = array([[vmap[k], vmap[k], 1.0] for k in self.variables])
+            A = csc_matrix(
+                (Av[:, 2], (Av[:, 0].astype("int64"), Av[:, 1].astype("int64"))),
+                shape=(nv, nv),
+            )
+            bounds = vbounds
         else:
-            self.status = "primal infeasible"
-            self.obj_value = None
-        self.info = solution.info
-        self.__solution = {
-            "x": solution.x,
-            "y": solution.y,
-            "rho": solution.info.rho_estimate,
-        }
+            A = None
+            bounds = None
+
+        integer = array([int(vn in self.integer_vars) for vn in self.variables])
+        return SparseProblem(P, q, A, bounds, vbounds, integer)
+
+    @abc.abstractmethod
+    def solve(self):
+        """Solve the problem.
+
+        This will return nothing but will fill the solution attributes.
+
+        Returns
+        -------
+        Nothing.
+        """
+        raise NotImplementedError("This needs to be overwritten by the child class.")
 
     def reset(self, everything=False):
-        """Reset the public solver solution."""
+        """Reset the public solver solution.
+
+        Parameters
+        ----------
+        everything : bool
+            If true will also clear cached solution bases or warm start solutions.
+
+        Returns
+        -------
+        Nothing.
+        """
         self.info = None
         self.primals = {}
         self.cprimals = {}
         self.duals = {}
         self.vduals = {}
         if everything:
-            self.__solution = None
+            self._solution = None
 
+    @abc.abstractmethod
     def still_valid(self, A, bounds):
-        """Check if previous solutions is still feasible."""
-        if len(self.__solution["x"]) != len(self.variables) or len(
-            self.__solution["y"]
-        ) != len(self.constraints):
-            return False
-        c = A.dot(self.__solution["x"])
-        ea = self.settings["eps_abs"]
-        er = self.settings["eps_rel"]
-        valid = np.all(
-            (c + er * np.abs(c) + ea >= bounds[:, 0])
-            & (c - er * np.abs(c) - ea <= bounds[:, 1])
-        )
-        return valid
+        """Check if previous solutions is still feasible.
 
-    def clean(self):
+        Parameters
+        ----------
+        A : numpy.csc_matrix
+            The constraint matrix A.
+        bounds : numpy.array
+            The constraint bounds.
+
+        Returns
+        -------
+        bool
+            Whether the current solution still fulfills the bounds.
+
+        """
+        raise NotImplementedError("This needs to be overwritten by the child class.")
+
+    def prune(self):
         """Remove unused variables and constraints."""
         self.reset()
         self.variable_lbs = {
-            k: v for k, v in six.iteritems(self.variable_lbs) if k in self.variables
+            k: v for k, v in self.variable_lbs.items() if k in self.variables
         }
         self.variable_ubs = {
-            k: v for k, v in six.iteritems(self.variable_ubs) if k in self.variables
+            k: v for k, v in self.variable_ubs.items() if k in self.variables
         }
         self.constraint_coefs = {
             k: v
-            for k, v in six.iteritems(self.constraint_coefs)
+            for k, v in self.constraint_coefs.items()
             if k[0] in self.constraints and k[1] in self.variables
         }
         self.obj_linear_coefs = {
-            k: v for k, v in six.iteritems(self.obj_linear_coefs) if k in self.variables
+            k: v for k, v in self.obj_linear_coefs.items() if k in self.variables
         }
         self.obj_quadratic_coefs = {
             k: v
-            for k, v in six.iteritems(self.obj_quadratic_coefs)
+            for k, v in self.obj_quadratic_coefs.items()
             if k[0] in self.variables and k[1] in self.variables
         }
         self.constraint_lbs = {
-            k: v for k, v in six.iteritems(self.constraint_lbs) if k in self.constraints
+            k: v for k, v in self.constraint_lbs.items() if k in self.constraints
         }
         self.constraint_ubs = {
-            k: v for k, v in six.iteritems(self.constraint_ubs) if k in self.constraints
+            k: v for k, v in self.constraint_ubs.items() if k in self.constraints
         }
+        self.integer_vars = {v for v in self.integer_vars if v in self.variables}
 
     def rename_constraint(self, old, new):
-        """Rename a constraint."""
+        """Rename a constraint.
+
+        Parameters
+        ----------
+        old : str
+            The old name of the constraint.
+        new : str
+            The new name of the constraint.
+
+        Returns
+        -------
+        Nothing.
+
+        """
         self.reset()
         self.constraints.remove(old)
         self.constraints.add(new)
         name_map = {k: k for k in self.constraints}
         name_map[old] = new
         self.constraint_coefs = {
-            (name_map[k[0]], k[1]): v for k, v in six.iteritems(self.constraint_coefs)
+            (name_map[k[0]], k[1]): v for k, v in self.constraint_coefs.items()
         }
         self.constraint_lbs[new] = self.constraint_lbs.pop(old)
         self.constraint_ubs[new] = self.constraint_ubs.pop(old)
 
     def rename_variable(self, old, new):
-        """Rename a variable."""
+        """Rename a variable.
+
+        Parameters
+        ----------
+        old : str
+            The old name of the variable.
+        new : str
+            The new name of the variable.
+
+        Returns
+        -------
+        Nothing.
+
+        """
         self.reset()
         self.variables.remove(old)
         self.variables.add(new)
         name_map = {k: k for k in self.variables}
         name_map[old] = new
         self.constraint_coefs = {
-            (k[0], name_map[k[1]]): v for k, v in six.iteritems(self.constraint_coefs)
+            (k[0], name_map[k[1]]): v for k, v in self.constraint_coefs.items()
         }
         self.obj_quadratic_coefs = {
             (name_map[k[0]], name_map[k[1]]): v
-            for k, v in six.iteritems(self.obj_quadratic_coefs)
+            for k, v in self.obj_quadratic_coefs.items()
         }
         self.obj_linear_coefs = {
-            name_map[k]: v for k, v in six.iteritems(self.obj_linear_coefs)
+            name_map[k]: v for k, v in self.obj_linear_coefs.items()
         }
         self.variable_lbs[new] = self.variable_lbs.pop(old)
         self.variable_ubs[new] = self.variable_ubs.pop(old)
+        if old in self.integer_vars:
+            self.integer_vars.remove(old)
+            self.integer_vars.add(new)
 
 
-@six.add_metaclass(inheritdocstring)
 class Variable(interface.Variable):
     def __init__(self, name, *args, **kwargs):
-        super(Variable, self).__init__(name, **kwargs)
+        super().__init__(name=name, **kwargs)
 
     @interface.Variable.type.setter
     def type(self, value):
         if self.problem is not None:
             if value not in _TYPES:
                 raise ValueError(
-                    "OSQP cannot handle variables of type '%s'. " % value
+                    "This interface cannot handle variables of type '%s'. " % value
                     + "The following variable types are available: "
                     + ", ".join(_TYPES)
                 )
+            if value == "binary":
+                self.lb = 0
+                self.ub = 1
+                self.problem.problem.integer_vars.add(self.name)
+            elif value == "integer":
+                self.problem.problem.integer_vars.add(self.name)
+            elif value == "continuous":
+                if self.name in self.problem.problem.integer_vars:
+                    self.problem.problem.integer_vars.remove(self.name)
         super(Variable, Variable).type.fset(self, value)
 
     def _get_primal(self):
@@ -320,6 +400,8 @@ class Variable(interface.Variable):
 
     @property
     def dual(self):
+        if self.problem.is_integer:
+            raise ValueError("Dual values are not well-defined for integer problems")
         if self.problem is not None:
             return self.problem.problem.vduals.get(self.name, None)
         return None
@@ -333,18 +415,17 @@ class Variable(interface.Variable):
                 self.problem.problem.rename_variable(old_name, value)
 
 
-@six.add_metaclass(inheritdocstring)
 class Constraint(interface.Constraint):
     _INDICATOR_CONSTRAINT_SUPPORT = False
 
-    def __init__(self, expression, sloppy=False, *args, **kwargs):
-        super(Constraint, self).__init__(expression, *args, sloppy=sloppy, **kwargs)
+    def __init__(self, expression, sloppy=False, **kwargs):
+        super().__init__(expression=expression, sloppy=sloppy, **kwargs)
 
     def set_linear_coefficients(self, coefficients):
         if self.problem is not None:
             self.problem.update()
             self.problem.problem.reset()
-            for var, coef in six.iteritems(coefficients):
+            for var, coef in coefficients.items():
                 self.problem.problem.constraint_coefs[(self.name, var.name)] = float(
                     coef
                 )
@@ -395,11 +476,11 @@ class Constraint(interface.Constraint):
 
     @property
     def dual(self):
+        if self.problem.is_integer:
+            raise ValueError("Dual values are not well-defined for integer problems")
         if self.problem is None:
             return None
         d = self.problem.problem.duals.get(self.name, None)
-        if d is not None:
-            d = -d
         return d
 
     @interface.Constraint.name.setter
@@ -432,20 +513,20 @@ class Constraint(interface.Constraint):
         if self.problem is not None:
             problem_reference = self.problem
             self.problem._remove_constraint(self)
-            super(Constraint, self).__iadd__(other)
+            super().__iadd__(other)
             problem_reference._add_constraint(self, sloppy=False)
         else:
-            super(Constraint, self).__iadd__(other)
+            super().__iadd__(other)
         return self
 
 
-@six.add_metaclass(inheritdocstring)
 class Objective(interface.Objective):
     def __init__(self, expression, sloppy=False, **kwargs):
-        super(Objective, self).__init__(expression, sloppy=sloppy, **kwargs)
+        super().__init__(expression=expression, sloppy=sloppy, **kwargs)
         self._expression_expired = False
         if not (sloppy or self.is_Linear or self.is_Quadratic):
-            raise ValueError("OSQP only supports linear and quadratic objectives.")
+            raise ValueError(
+                "This interface only supports linear and quadratic objectives.")
 
     @property
     def value(self):
@@ -474,13 +555,13 @@ class Objective(interface.Objective):
             expression = add(
                 [
                     coef * vars[vn]
-                    for vn, coef in six.iteritems(model.problem.obj_linear_coefs)
+                    for vn, coef in model.problem.obj_linear_coefs.items()
                 ]
             )
             q_ex = add(
                 [
                     coef * vars[vn[0]] * vars[vn[1]]
-                    for vn, coef in six.iteritems(model.problem.obj_quadratic_coefs)
+                    for vn, coef in model.problem.obj_quadratic_coefs.items()
                 ]
             )
             expression += q_ex
@@ -491,7 +572,7 @@ class Objective(interface.Objective):
     def set_linear_coefficients(self, coefficients):
         if self.problem is not None:
             self.problem.update()
-            for v, coef in six.iteritems(coefficients):
+            for v, coef in coefficients.items():
                 self.problem.problem.obj_linear_coefs[v.name] = float(coef)
             self._expression_expired = True
         else:
@@ -514,64 +595,50 @@ class Objective(interface.Objective):
             )
 
 
-@six.add_metaclass(inheritdocstring)
 class Configuration(interface.MathematicalProgrammingConfiguration):
+    lp_methods = _LP_METHODS
+    qp_methods = _QP_METHODS
+
     def __init__(
         self,
-        lp_method="primal",
+        lp_method="auto",
         presolve=False,
         verbosity=0,
         timeout=None,
-        qp_method="primal",
-        linear_solver="qdldl",
-        *args,
+        qp_method="auto",
         **kwargs
     ):
-        super(Configuration, self).__init__(*args, **kwargs)
+        super().__init__(**kwargs)
         self.lp_method = lp_method
         self.presolve = presolve
         self.verbosity = verbosity
         self.timeout = timeout
         self.qp_method = qp_method
-        self.linear_solver = linear_solver
         if "tolerances" in kwargs:
-            for key, val in six.iteritems(kwargs["tolerances"]):
+            for key, val in kwargs["tolerances"].items():
                 if key in self._tolerance_functions():
                     setattr(self.tolerances, key, val)
 
     @property
     def lp_method(self):
         """The algorithm used to solve LP problems."""
-        return "primal"
+        return self.problem.problem.settings["lp_method"]
 
     @lp_method.setter
-    def lp_method(self, lp_method):
-        if lp_method not in _LP_METHODS:
+    def lp_method(self, value):
+        if value not in self.lp_methods:
             raise ValueError(
                 "LP Method %s is not valid (choose one of: %s)"
-                % (lp_method, ", ".join(_LP_METHODS))
-            )
-
-    @property
-    def linear_solver(self):
-        return self._linear_solver
-
-    @linear_solver.setter
-    def linear_solver(self, solver):
-        if solver not in ("qdldl", "mkl pardiso"):
-            raise ValueError(
-                "%s is not valid (choose either `qdldl` or `mkl pardiso`)" % solver
+                % (value, ", ".join(self.lp_methods))
             )
         if getattr(self, "problem", None) is not None:
-            self.problem.problem.settings["linsys_solver"] = solver
-        self._linear_solver = solver
+            self.problem.problem.settings["lp_method"] = value
+        self._lp_method = value
 
     def _set_presolve(self, value):
         if getattr(self, "problem", None) is not None:
-            if value is True:
-                self.problem.problem.settings["scaling"] = 10
-            elif value is False or value == "auto":
-                self.problem.problem.settings["scaling"] = 0
+            if value in [True, False, "auto"]:
+                self.problem.problem.settings["presolve"] = value
             else:
                 raise ValueError(
                     str(value) + " is not a valid presolve parameter. Must be True, "
@@ -580,7 +647,7 @@ class Configuration(interface.MathematicalProgrammingConfiguration):
 
     @property
     def presolve(self):
-        return self.problem.problem.settings["scaling"] > 0
+        return self.problem.problem.settings["presolve"]
 
     @presolve.setter
     def presolve(self, value):
@@ -594,7 +661,7 @@ class Configuration(interface.MathematicalProgrammingConfiguration):
     @verbosity.setter
     def verbosity(self, value):
         if getattr(self, "problem", None) is not None:
-            self.problem.problem.settings["verbose"] = int(value > 0)
+            self.problem.problem.settings["verbose"] = int(value)
         self._verbosity = value
 
     @property
@@ -614,7 +681,6 @@ class Configuration(interface.MathematicalProgrammingConfiguration):
             "presolve": self.presolve,
             "timeout": self.timeout,
             "verbosity": self.verbosity,
-            "linear_solver": self.linear_solver,
             "tolerances": {
                 "feasibility": self.tolerances.feasibility,
                 "optimality": self.tolerances.optimality,
@@ -623,46 +689,47 @@ class Configuration(interface.MathematicalProgrammingConfiguration):
         }
 
     def __setstate__(self, state):
-        for key, val in six.iteritems(state):
+        for key, val in state.items():
             if key != "tolerances":
                 setattr(self, key, val)
-        for key, val in six.iteritems(state["tolerances"]):
+        for key, val in state["tolerances"].items():
             if key in self._tolerance_functions():
                 setattr(self.tolerances, key, val)
 
     @property
     def qp_method(self):
         """Change the algorithm used to optimize QP problems."""
-        return "primal"
+        return self.problem.problem.settings["qp_method"]
 
     @qp_method.setter
     def qp_method(self, value):
         if value not in _QP_METHODS:
             raise ValueError(
                 "%s is not a valid qp_method. Choose between %s"
-                % (value, str(_QP_METHODS))
+                % (value, str(self.qp_methods))
             )
+        if getattr(self, "problem", None) is not None:
+            self.problem.problem.settings["qp_method"] = value
         self._qp_method = value
 
     def _get_feasibility(self):
-        return self.problem.problem.settings["eps_prim_inf"]
+        return self.problem.problem.settings["primal_inf_tolerance"]
 
     def _set_feasibility(self, value):
-        self.problem.problem.settings["eps_prim_inf"] = value
-        self.problem.problem.settings["eps_dual_inf"] = value
+        self.problem.problem.settings["primal_inf_tolerance"] = value
+        self.problem.problem.settings["dual_inf_tolerance"] = value
 
     def _get_integrality(self):
-        return 1e-6
+        return self.problem.problem.settings["mip_tolerance"]
 
     def _set_integrality(self, value):
-        pass
+        self.problem.problem.settings["mip_tolerance"] = value
 
     def _get_optimality(self):
-        return self.problem.problem.settings["eps_abs"]
+        return self.problem.problem.settings["optimality_tolerance"]
 
     def _set_optimality(self, value):
-        self.problem.problem.settings["eps_abs"] = value
-        self.problem.problem.settings["eps_rel"] = value
+        self.problem.problem.settings["optimality_tolerance"] = value
 
     def _tolerance_functions(self):
         return {
@@ -672,13 +739,15 @@ class Configuration(interface.MathematicalProgrammingConfiguration):
         }
 
 
-@six.add_metaclass(inheritdocstring)
 class Model(interface.Model):
+    ProblemClass = MatrixProblem
+    status_map = _STATUS_MAP
+
     def _initialize_problem(self):
-        self.problem = OSQPProblem()
+        self.problem = self.ProblemClass()
 
     def _initialize_model_from_problem(self, problem, vc_mapping=None, offset=0):
-        if not isinstance(problem, OSQPProblem):
+        if not isinstance(problem, self.ProblemClass):
             raise TypeError("Provided problem is not a valid OSQP model.")
         self.problem = problem
         for name in self.problem.variables:
@@ -688,7 +757,7 @@ class Model(interface.Model):
                 ub=self.problem.variable_ubs[name],
                 problem=self,
             )
-            super(Model, self)._add_variables([var])
+            super()._add_variables([var])
 
         for name in self.problem.constraints:
             # Since constraint expressions are lazily retrieved from the
@@ -702,7 +771,7 @@ class Model(interface.Model):
                 problem=self,
             )
 
-            super(Model, self)._add_constraints([constr], sloppy=True)
+            super()._add_constraints([constr], sloppy=True)
 
         if vc_mapping is None:
             for constr in self.constraints:
@@ -720,13 +789,13 @@ class Model(interface.Model):
         linear_expression = add(
             [
                 coef * self._variables[vn]
-                for vn, coef in six.iteritems(self.problem.obj_linear_coefs)
+                for vn, coef in self.problem.obj_linear_coefs.items()
             ]
         )
         quadratic_expression = add(
             [
                 coef * self._variables[vn[0]] * self._variables[vn[1]]
-                for vn, coef in six.iteritems(self.problem.obj_quadratic_coefs)
+                for vn, coef in self.problem.obj_quadratic_coefs.items()
             ]
         )
 
@@ -735,7 +804,7 @@ class Model(interface.Model):
             linear_expression + quadratic_expression + offset,
             problem=self,
             direction={-1: "max", 1: "min"}[self.problem.direction],
-            name="osqp_objective",
+            name="matrix_objective",
         )
 
     @property
@@ -759,12 +828,12 @@ class Model(interface.Model):
         self._objective_offset = offset
         if linear_coefficients:
             self.problem.obj_linear_coefs = {
-                v.name: float(c) for v, c in six.iteritems(linear_coefficients)
+                v.name: float(c) for v, c in linear_coefficients.items()
             }
 
-        for key, coef in six.iteritems(quadratic_coeffients):
+        for key, coef in quadratic_coeffients.items():
             if len(key) == 1:
-                var = six.next(iter(key))
+                var = next(iter(key))
                 self.problem.obj_quadratic_coefs[(var.name, var.name)] = float(coef)
             else:
                 var1, var2 = key
@@ -788,33 +857,43 @@ class Model(interface.Model):
         return primal_values
 
     def _get_reduced_costs(self):
-        if len(self.problem.duals) == 0:
+        if self.is_integer:
+            raise ValueError("Dual values are not well-defined for integer problems")
+        if len(self.problem.vduals) == 0:
             raise SolverError("The problem has not been solved yet!")
         reduced_costs = [self.problem.vduals[v.name] for v in self._variables]
         return reduced_costs
 
     def _get_constraint_values(self):
-        if len(self.problem.primals) == 0:
+        if len(self.problem.cprimals) == 0:
             raise SolverError("The problem has not been solved yet!")
         constraint_primals = [self.problem.cprimals[c.name] for c in self._constraints]
         return constraint_primals
 
     def _get_shadow_prices(self):
-        if len(self.problem.primals) == 0:
+        if self.is_integer:
+            raise ValueError("Dual values are not well-defined for integer problems")
+        if len(self.problem.duals) == 0:
             raise SolverError("The problem has not been solved yet!")
-        dual_values = [-self.problem.duals[c.name] for c in self._constraints]
+        dual_values = [self.problem.duals[c.name] for c in self._constraints]
         return dual_values
 
     @property
     def is_integer(self):
-        return False
+        return len(self.problem.integer_vars) > 0
 
     def _optimize(self):
         self.update()
         self.problem.solve()
-        osqp_status = self.problem.status
-        self._original_status = osqp_status
-        status = _STATUS_MAP[osqp_status]
+        prior_status = self.problem.status
+        self._original_status = prior_status
+        status = self.status_map[prior_status]
+        return status
+
+    def optimize(self):
+        self.update()
+        status = self._optimize()
+        self._status = status
         return status
 
     def _set_variable_bounds_on_problem(self, var_lb, var_ub):
@@ -827,7 +906,7 @@ class Model(interface.Model):
             self.problem.variable_ubs[var.name] = float(ub)
 
     def _add_variables(self, variables):
-        super(Model, self)._add_variables(variables)
+        super()._add_variables(variables)
         self.problem.reset()
         for variable in variables:
             lb = -Infinity if variable.lb is None else float(variable.lb)
@@ -835,10 +914,13 @@ class Model(interface.Model):
             self.problem.variables.add(variable.name)
             self.problem.variable_lbs[variable.name] = lb
             self.problem.variable_ubs[variable.name] = ub
+            if variable.type in ("binary", "integer"):
+                self.problem.integer_vars.add(variable.name)
             variable.problem = self
 
     def _remove_variables(self, variables):
-        # Not calling parent method to avoid expensive variable removal from sympy expressions
+        # Not calling parent method to avoid expensive variable removal
+        # from sympy expressions
         if self.objective is not None:
             self.objective._expression = self.objective.expression.xreplace(
                 {var: 0 for var in variables}
@@ -848,13 +930,14 @@ class Model(interface.Model):
             variable.problem = None
             del self._variables[variable.name]
             self.problem.variables.remove(variable.name)
-        self.problem.clean()
+        self.problem.prune()
 
     def _add_constraints(self, constraints, sloppy=False):
-        super(Model, self)._add_constraints(constraints, sloppy=sloppy)
+        super()._add_constraints(constraints, sloppy=sloppy)
         self.problem.reset()
         for constraint in constraints:
-            constraint._problem = None  # This needs to be done in order to not trigger constraint._get_expression()
+            # This needs to be done in order to not trigger constraint._get_expression()
+            constraint._problem = None
             if constraint.is_Linear:
                 _, coeff_dict, _ = parse_optimization_expression(constraint)
                 lb = -Infinity if constraint.lb is None else float(constraint.lb)
@@ -863,7 +946,7 @@ class Model(interface.Model):
                 self.problem.constraint_coefs.update(
                     {
                         (constraint.name, v.name): float(co)
-                        for v, co in six.iteritems(coeff_dict)
+                        for v, co in coeff_dict.items()
                     }
                 )
                 self.problem.constraint_lbs[constraint.name] = lb
@@ -872,19 +955,19 @@ class Model(interface.Model):
             elif constraint.is_Quadratic:
                 raise NotImplementedError(
                     "Quadratic constraints (like %s) are not supported "
-                    "in OSQP yet." % constraint
+                    "in this interface yet." % constraint
                 )
             else:
                 raise ValueError(
-                    "OSQP only supports linear or quadratic constraints. "
+                    "This interface only supports linear or quadratic constraints. "
                     "%s is neither linear nor quadratic." % constraint
                 )
 
     def _remove_constraints(self, constraints):
-        super(Model, self)._remove_constraints(constraints)
+        super()._remove_constraints(constraints)
         for constraint in constraints:
             self.problem.constraints.remove(constraint.name)
-        self.problem.clean()
+        self.problem.prune()
 
     def _get_variable_indices(self, names):
         vmap = dict(zip(self.variables, range(len(self.variables))))
@@ -892,16 +975,16 @@ class Model(interface.Model):
 
     def __setstate__(self, d):
         self.__init__()
-        osqp = pickle.loads(d["osqp"])
+        problem = pickle.loads(d["problem"])
         # workaround to conserve the order
-        osqp.variables = d["variables"]
-        osqp.constraints = d["constraints"]
+        problem.variables = d["variables"]
+        problem.constraints = d["constraints"]
         self._initialize_model_from_problem(
-            osqp, vc_mapping=d["v_to_c"], offset=d["offset"]
+            problem, vc_mapping=d["v_to_c"], offset=d["offset"]
         )
-        osqp.variables = set(osqp.variables)
-        osqp.constraints = set(osqp.constraints)
-        self.configuration = Configuration()
+        problem.variables = set(problem.variables)
+        problem.constraints = set(problem.constraints)
+        self.configuration = self.interface.Configuration()
         self.configuration.problem = self
         self.configuration.__setstate__(d["config"])
 
@@ -909,7 +992,7 @@ class Model(interface.Model):
         self.problem.reset()
         self.update()
         return {
-            "osqp": pickle.dumps(self.problem),
+            "problem": pickle.dumps(self.problem),
             "variables": [v.name for v in self._variables],
             "constraints": [c.name for c in self._constraints],
             "v_to_c": self._variables_to_constraints_mapping,
@@ -921,7 +1004,7 @@ class Model(interface.Model):
     def from_lp(self, lp_problem_str):
         """Read a model from an LP file.
 
-        OSQP does not have an integrated LP reader so it will either use
+        The solver may not have an integrated LP reader so it will either use
         cplex or glpk to read the model. This means that QP problems will
         currently require cplex to be read :(
         """
@@ -931,10 +1014,10 @@ class Model(interface.Model):
             mod = cplex_interface.Model.from_lp(lp_problem_str)
             mod.configuration.lp_method = "auto"
             mod.configuration.qp_method = "auto"
-            return super(Model, self).clone(mod)
+            return super().clone(mod)
         else:
             from optlang import glpk_interface
 
             mod = glpk_interface.Model.from_lp(lp_problem_str)
             mod.configuration.lp_method = "auto"
-            return super(Model, self).clone(mod)
+            return super().clone(mod)